requirements.txt
boto3
webauthn
itsdangerous
python-jose
requests
pydantic
pydantic-settings